requirements.txt
pyyaml>=6.0.1
hatchling>=1.18.0  # For building the package 